refactor: query client key helper
